requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
resend==0.6.0
requests==2.31.0

